HDFS-16174. Implement temp file and dir in cc files
* Need to implement TempFile and
  TempDir in cc files for avoiding
  duplicate compilation of these
  classes.
